modified login

+++ coffee-frontend/coffee-frontend/src/app/components/layout/layout.ts
@@ -1,7 +1,8 @@
-import { Component, inject } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { Router, RouterModule, RouterOutlet } from '@angular/router';
 import { IAuthResponse } from '../../models/auth.model';
 import { CommonModule } from '@angular/common';
+import { AuthService } from '../../services/auth/auth';
 
 @Component({
   selector: 'app-layout',
@@ -9,15 +10,37 @@ import { CommonModule } from '@angular/common';
   templateUrl: './layout.html',
   styleUrl: './layout.css'
 })
-export class Layout {
+export class Layout implements OnInit {
 
   router = inject(Router)
   user! :IAuthResponse;
+  isLoggedIn = true;
+
+  authService = inject(AuthService)
+  userObj: { email: string } = { email: 'dadad' };
+
+
+  
+
+  ngOnInit(): void {
+    this.authService.isLoggedIn$.subscribe((status)=>{
+      this.isLoggedIn = status;
+    })
+    console.log(this.userObj.email)
+  }
+  
 
 
   onLogout(){
-    localStorage.clear(); // clears all localStorage data
-  this.router.navigate(['/login']);
+    const logout = confirm("Are you sure you want to  logout?")
+    if(logout){
+      localStorage.clear(); // clears all localStorage data
+      this.authService['loggedIn'].next(false)
+      this.router.navigate(['/login']);
+      this.isLoggedIn = false;
+
+    }
+    
 
   }
 }
